requirements.txt
Django==1.4
South==0.7.5
django-crispy-forms==1.1.4
django-floppyforms==0.4.7
django-social-auth==0.6.9
django-taggit==0.9.3
httplib2==0.7.4
mock==0.8.0
oauth2==1.5.211
psycopg2==2.4.5
python-openid==2.2.5
wsgiref==0.1.2
